Pipe configured polling place ID through to subocomponents
Following the pattern set for the precinct selection for now, where
`AppRoot` fetches the election config data and passes that down into
child screens/subcomponents. Could be later improved by moving the
relevant data queries into components that need them/creating helper
hooks to handle optional vs required configuration.

Aso updating the `UnconfiguredPrecinctScreen` to
`UnconfiguredPollingPlaceScreen`, with a temporary conditional render
based on the `ENABLE_POLLING_PLACES` flag.
