+++ apps/scan/frontend/src/app_root.tsx
@@ -8,8 +8,10 @@ import {
   Keybinding,
 } from '@votingworks/ui';
 import {
+  BooleanEnvironmentVariableName as Feature,
   isSystemAdministratorAuth,
   isElectionManagerAuth,
+  isFeatureFlagEnabled,
   isPollWorkerAuth,
   isVendorAuth,
 } from '@votingworks/utils';
@@ -25,7 +27,7 @@ import { CardErrorScreen } from './screens/card_error_screen';
 import { InternalConnectionProblemScreen } from './screens/internal_connection_problem_screen';
 import { InsertUsbScreen } from './screens/insert_usb_screen';
 import { POLLING_INTERVAL_FOR_SCANNER_STATUS_MS } from './config/globals';
-import { UnconfiguredPrecinctScreen } from './screens/unconfigured_precinct_screen';
+import { UnconfiguredPollingPlaceScreen } from './screens/unconfigured_polling_place_screen';
 import { UnconfiguredElectionScreenWrapper } from './screens/unconfigured_election_screen_wrapper';
 import {
   checkPin,
@@ -176,6 +178,7 @@ export function AppRoot(): JSX.Element | null {
     systemSettings,
     isTestMode,
     precinctSelection,
+    pollingPlaceId,
     isSoundMuted,
     isContinuousExportEnabled,
     ballotCastingMode,
@@ -331,7 +334,13 @@ export function AppRoot(): JSX.Element | null {
     );
   }
 
-  if (!precinctSelection) return <UnconfiguredPrecinctScreen />;
+  const locationConfigured = isFeatureFlagEnabled(Feature.ENABLE_POLLING_PLACES)
+    ? !!pollingPlaceId
+    : !!precinctSelection;
+
+  if (!locationConfigured) {
+    return <UnconfiguredPollingPlaceScreen />;
+  }
 
   if (isPollWorkerAuth(authStatus)) {
     return (
