--- apps/scan/frontend/src/screens/unconfigured_precinct_screen.tsx
@@ -1,22 +1,0 @@
-import { H1, P } from '@votingworks/ui';
-import { CenteredText, ScreenMainCenterChild } from '../components/layout';
-
-export function UnconfiguredPrecinctScreen(): JSX.Element {
-  return (
-    <ScreenMainCenterChild
-      voterFacing={false}
-      showTestModeBanner={false}
-      showEarlyVotingBanner={false}
-    >
-      <CenteredText>
-        <H1>No Precinct Selected</H1>
-        <P>Insert an election manager card to select a precinct.</P>
-      </CenteredText>
-    </ScreenMainCenterChild>
-  );
-}
-
-/* istanbul ignore next - @preserve */
-export function DefaultPreview(): JSX.Element {
-  return <UnconfiguredPrecinctScreen />;
-}
